// src/lib/types.ts
export interface FontUsage {
  fontWeight: string;
  fontStyle: string;
  fontStretch: string;
  fontSize: string;
}

export interface FontFaceData {
  stretch: string;
  style: string;
  weight: string;
}

export interface FontData {
  familyName: string;
  faces: FontFaceData[];
  usages: FontUsage[];
}